refactor(mcp): tidy src layout, drop semantic-layer-mcp shim, harden playbook
Remove orphaned pre-persona modules and the unpublished semantic-layer-mcp package, fold runtime config/audit helpers, and revise semantic-layer prompts/playbook from live tool audits.
