+++ .changes/unreleased/refactor-20260730-220741.yaml
@@ -1,0 +1,3 @@
+kind: refactor
+body: Remove orphaned pre-persona MCP modules (package-level prompts/resources/completion/tasks/ai-agents tools) and fold runtime config + audit helpers under config/ and audit/
+time: 2026-07-30T22:07:41.446161+09:00
