+++ AGENTS.md
@@ -178,7 +178,7 @@ Additional specialized skills are documented in `CLAUDE.md`.
 - **Stdio MCP smoke env:** `EnvContextProvider` calls `getClient()` at construct time, so `LIGHTDASH_URL` and `LIGHTDASH_API_KEY` must be set before the process starts. Dummy values are enough for initialize/tools/list; the Lightdash API is only hit on tool execution.
 - **OpenAPI pin lives in `config/lightdash-openapi-ref.txt`:** `generate:types` fetches swagger from that commit SHA (not floating `main`). Bump the pin, regenerate, then fix breakages — never hand-edit `openapi-types.ts`.
 - **`LightdashUser` / `AuthenticatedUser` avatar fields:** Upstream now requires `avatarUrl` and `avatarGradient` (`string | null`). Fixtures using `satisfies AuthenticatedUser` must include both after an OpenAPI sync.
-- **MCP personas + shared tool registry (ADR-0042):** Shared tools live in `packages/mcp/src/tools/registry.ts`; personas under `packages/mcp/src/personas/<id>/` own `toolIds`, prompts, playbooks, and a fixed HTTP `path`. Shipped persona is `semantic-layer` only (`/semantic-layer/v1/mcp`, nine `ldt__*` tools via `registerToolSafe`). No `LIGHTDASH_TOOLS_MCP_PERSONA` / free-form `LIGHTDASH_TOOLS_MCP_PATH`. Stdio always uses the sole shipped persona. Broad admin ops stay on client/CLI. `@lightdash-tools/semantic-layer-mcp` is a deprecated shim.
+- **MCP personas + shared tool registry (ADR-0042):** Shared tools live in `packages/mcp/src/tools/registry.ts`; personas under `packages/mcp/src/personas/<id>/` own `toolIds`, prompts, playbooks, and a fixed HTTP `path`. Shipped persona is `semantic-layer` only (`/semantic-layer/v1/mcp`, nine `ldt__*` tools via `registerToolSafe`). No `LIGHTDASH_TOOLS_MCP_PERSONA` / free-form `LIGHTDASH_TOOLS_MCP_PATH`. Stdio always uses the sole shipped persona. Broad admin ops stay on client/CLI. Do not recreate `@lightdash-tools/semantic-layer-mcp` — that package was removed. Runtime client/guardrails live in `packages/mcp/src/config/runtime.ts`; audit helpers in `packages/mcp/src/audit/`. Package-level `src/prompts/` / `src/resources/` were removed — prompts/resources are persona-owned only.
 - **Semantic-layer MCP UX:** Prompts (`lightdash_semantic_explore`, `lightdash_semantic_compose_compile`, `lightdash_semantic_compile_debug`) + playbook `lightdash://playbooks/semantic-layer` cite `ldt__*` names. `list_explores` returns summaries with optional `search`/`limit`; `list_dimensions` includes `fieldId` for `compile_query`. Compile/discovery only — no run-query. Empty SELECT → `isError`. Optional project pin is HTTP-only via `X-Lightdash-Project`.
 - **MCP Docker / Compose:** Build from repo root: `docker build -f packages/mcp/Dockerfile -t lightdash-mcp:local .`. Root `.dockerignore` → `packages/mcp/.dockerignore`. Compose: `docker compose -f docker-compose.dev.yml --profile semantic-layer up --build` → Cursor `url: http://localhost:8080/semantic-layer/v1/mcp` with `AUTH_MODE=none` + `.env` PAT.
