+++ packages/mcp/src/audit/audit.ts
