+++ packages/mcp/src/audit/tool-audit-context.ts
