+++ packages/mcp/src/personas/semantic-layer/prompts.ts
@@ -64,9 +64,9 @@ export function registerSemanticLayerPrompts(server: McpServer): void {
           `Project: ${projectUuid}.${topic}`,
           SEMANTIC_LAYER_HARD_BANS,
           'Procedure:',
-          '1) Map any BigQuery/warehouse hints (dataset.table) → list_explores search on the table token.',
-          '2) Disambiguate: exact label match + schemaName (prefer dwh_* warehouse table over eda_/reporting_/mre_ unless asked). Never pick the first hit blindly.',
-          '3) list_dimensions (base-table fieldIds) and/or list_metrics with metric keywords from the question (nps, count, …) — not the table name. Filter metrics where tableName === explore id.',
+          '1) If the user gave BigQuery/`bq ls` hints: dataset → prefer schemaName; TABLE → list_explores search; partition/time field → candidate dimension later.',
+          '2) Always ldt__list_explores with search+limit (never default first-N). Disambiguate: exact label + schemaName (prefer dwh_* warehouse table; skip empty schemaName when a dataset was given; deprioritize eda_/reporting_/mre_).',
+          '3) ldt__list_dimensions (base-table fieldIds). For metrics: search specific keywords from the question (nps, session, …) — not the table/explore id; avoid broad-only tokens like count/sum as a first search. Keep rows with tableName === explore id.',
           'Deliverable: shortlist name/label/schemaName/tags/fieldId (+ why chosen). Do not dump get_explore or full dimension arrays.',
           `Follow the embedded playbook (${SEMANTIC_LAYER_PLAYBOOK_URI}). Use only ldt__* tools.`,
         ].join('\n'),
@@ -95,8 +95,8 @@ export function registerSemanticLayerPrompts(server: McpServer): void {
           `Project: ${projectUuid}.${topic}${explore}`,
           SEMANTIC_LAYER_HARD_BANS,
           'Procedure:',
-          '1) If exploreId is missing, discover + disambiguate (exact label + schemaName; warehouse table beats eda_/reporting_/mre_).',
-          '2) Take dimension fieldIds from ldt__list_dimensions (base table). For metrics, search keywords from the goal (not table names), keep rows with tableName === explore id; metrics may be [].',
+          '1) If exploreId is missing, map any BigQuery/`bq ls` hints (dataset→schemaName, TABLE→search), then disambiguate (exact label + schemaName; warehouse table beats eda_/reporting_/mre_; skip empty schemaName when dataset known).',
+          '2) Dimension fieldIds from ldt__list_dimensions (base table). Metrics: keyword search → filter tableName === explore id; metric wire ids are {exploreId}_{metricName}. metrics may be [] (e.g. volume-by-day via a time dimension).',
           '3) ldt__compile_query with fieldIds only. If isError about empty SELECT, switch to fieldIds and re-compile once.',
           '4) Verify SQL columns match the goal; stop.',
           'Deliverable: compiled SQL (or compile errors). Never paste full explore/dimension payloads.',
@@ -128,7 +128,7 @@ export function registerSemanticLayerPrompts(server: McpServer): void {
           'Checklist:',
           '- Empty SELECT / isError about empty SELECT → replace short names with fieldIds from ldt__list_dimensions (base table).',
           '- Unknown field id → re-list dimensions; confirm explore id is the disambiguated warehouse match (exact label + schemaName).',
-          '- Metric not found → get_metric/list_metrics tableName must be the full explore id; search metric keywords not table names.',
+          '- Metric not found → get_metric/list_metrics tableName must be the full explore id; search metric keywords not table names; compile metrics as {exploreId}_{metricName}.',
           'Deliverable: successful compile (SQL with expected columns) or a clear blocker. Stop after a good compile.',
           `Follow the embedded playbook (${SEMANTIC_LAYER_PLAYBOOK_URI}).`,
         ].join('\n'),
