+++ packages/mcp/src/tools/shared.ts
@@ -14,7 +14,6 @@ import {
   isAllowed,
   areAllProjectsAllowed,
   extractProjectUuids,
-  hasYamlProjectDocumentArgs,
   READ_ONLY_DEFAULT,
   logAuditEntry,
   getSessionId,
@@ -22,6 +21,7 @@ import {
   validateResourceIdsInObject,
 } from '@lightdash-tools/common';
 
+import { getToolAuditAuth, runWithToolAuditAuthAsync } from '../audit/tool-audit-context.js';
 import {
   isReadOnlyMcpScope,
   RequiredMcpScope,
@@ -33,9 +33,8 @@ import {
   getSafetyMode,
   getAllowedProjectUuids,
   isDryRunMode,
-} from '../config.js';
+} from '../config/runtime.js';
 import { toMcpErrorMessage } from '../errors.js';
-import { getToolAuditAuth, runWithToolAuditAuthAsync } from '../tool-audit-context.js';
 
 import type { McpContextProvider } from '../request-context.js';
 import type { LightdashClient } from '@lightdash-tools/client';
@@ -270,18 +269,6 @@ export function registerToolSafe(
     const innerHandler = finalHandler;
     finalHandler = async (args, extra): Promise<TextContent> => {
       const projectUuids = extractProjectUuids(args);
-      if (hasYamlProjectDocumentArgs(args) && projectUuids.length === 0) {
-        return {
-          content: [
-            {
-              type: 'text',
-              text: 'Error: Could not extract project UUID from YAML document for allowlist check. Fix bundleYaml/gateYaml or omit the document.',
-            },
-          ],
-          isError: true,
-          _lightdashBlocked: true,
-        } as BlockedContent;
-      }
       const deniedUuids = projectUuids.filter(
         (uuid) => !areAllProjectsAllowed(allowedProjects, [uuid]),
       );
