--- packages/semantic-layer-mcp/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@lightdash-tools/semantic-layer-mcp",
-  "version": "0.6.0",
-  "description": "DEPRECATED: use @lightdash-tools/mcp (semantic-layer persona at /semantic-layer/v1/mcp).",
-  "keywords": [],
-  "license": "Apache-2.0",
-  "author": "",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "bin": {
-    "lightdash-semantic-layer-mcp": "./dist/bin.js"
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/bin.js",
-    "test": "pnpm --dir ../.. exec vitest run packages/semantic-layer-mcp/src"
-  },
-  "dependencies": {
-    "@lightdash-tools/mcp": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/node": "^25.9.3"
-  },
-  "deprecated": "Use @lightdash-tools/mcp instead. Semantic-layer tools ship as the default persona at /semantic-layer/v1/mcp with ldt__* names."
-}
