--- packages/semantic-layer-mcp/src/index.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated Use `@lightdash-tools/mcp` instead.
- * The semantic-layer persona lives at `/semantic-layer/v1/mcp` with `ldt__*` tools.
- */
-export {};
