--- packages/semantic-layer-mcp/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2017",
-    "module": "commonjs",
-    "declaration": true,
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "types": ["node"],
-    "ignoreDeprecations": "6.0"
-  },
-  "include": ["src"]
-}
