+++ vitest.config.ts
@@ -34,8 +34,6 @@ export default defineConfig({
         'packages/mcp/src/bin.ts',
         'packages/mcp/src/index.ts',
         'packages/mcp/src/http.ts',
-        'packages/semantic-layer-mcp/src/bin.ts',
-        'packages/semantic-layer-mcp/src/index.ts',
         'packages/cli/src/index.ts',
         'packages/cli/src/commands/agentops/**',
       ],
